+++ packages/wonder-blocks-testing-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @khanacademy/wonder-blocks-testing-core
 
+## 3.0.1
+
+### Patch Changes
+
+- 30f357a: Deprecate fixtures framework
+
 ## 3.0.0
 
 ### Major Changes